Working on implementing a Java interface for the API
