+++ converter/src/main/scala/Main.scala
@@ -1,10 +1,9 @@
 import java.io.File
 
-import ReaderConverter.conversion.FileType
-import ReaderConverter.conversion.Converter._
+import parser.conversion.FileType
+import parser.conversion.Converter._
 
 object Main {
-
 
   def main(args: Array[String]): Unit = {
     val filepath = new File("./converter/resources/cv.pdf")
