--- converter/src/main/scala/ReaderConverter/conversion/FileType.scala
@@ -1,11 +1,0 @@
-package ReaderConverter.conversion
-
-/**
-  * Enumeration for the supported file type conversions in reader.Converter.scala convertPDFtoIMG
-  */
-object FileType extends Enumeration {
-  val png: FileType.Value = Value("png")
-  val jpg: FileType.Value = Value("jpg")
-  val jpeg: FileType.Value = Value("jpeg")
-  val gif: FileType.Value = Value("gif")
-}
