+++ reader/src/main/scala/Main.scala
@@ -2,11 +2,11 @@ import java.io.File
 
 object Main {
 
-  import ReaderConverter.Extraction.Extractor._
+  import parser.extraction.Extractor._
 
   def main(args: Array[String]): Unit = {
     val filepath = new File("./reader/resources/MegaTester.pdf")
-    //    val text = readPDF(filepath)
+    val text = readPDF(filepath)
     //    getJSONObjects(text, List(("name", POSTag.PN), ("weight", POSTag.NUM)), "null").foreach(println)
     //        val objs = getAllObjects(text, List(("name", POSTag.PN), ("weight", POSTag.NUM)))
     //    println(objs)
